+++ server/src/weaverbird/pipeline/steps/filter.py
@@ -3,13 +3,16 @@ from typing import Literal
 from weaverbird.pipeline.steps.utils.base import BaseStep
 from weaverbird.pipeline.steps.utils.render_variables import StepWithVariablesMixin
 
-from ..conditions import Condition
+from ..conditions import Condition, ConditionWithVariables
 
 
-class FilterStep(BaseStep):
+class BaseFilterStep(BaseStep):
     name: Literal["filter"] = "filter"
+
+
+class FilterStep(BaseFilterStep):
     condition: Condition
 
 
-class FilterStepWithVariables(FilterStep, StepWithVariablesMixin):
-    ...
+class FilterStepWithVariables(BaseFilterStep, StepWithVariablesMixin):
+    condition: ConditionWithVariables